+++ tests/test_snaskool.py
@@ -2508,6 +2508,23 @@ class SkoolWriterTest(SkoolKitTestCase):
         snapshot = [175, 201]
         self._test_write_skool(snapshot, ctl, exp_skool)
 
+    def test_rom_directives_with_parameters(self):
+        ctl = """
+            @ 00000 rom=48
+            c 00000 Routine at 0
+            @ 00001 rom=128,1
+            i 00002
+        """
+        exp_skool = """
+            @rom=48
+            ; Routine at 0
+            c00000 XOR A         ;
+            @rom=128,1
+             00001 RET           ;
+        """
+        snapshot = [175, 201]
+        self._test_write_skool(snapshot, ctl, exp_skool)
+
     def test_set_directives(self):
         ctl = """
             @ 00000 set-crlf=1